Cache the whole setup

+++ R/setup_travis.R
@@ -4,9 +4,17 @@
 #' @export
 
 setup_travis <- function() {
-  system("source ./travis/texlive.sh")
-  system("tlmgr install acronym bigfoot blindtext chngcntr cmap nowidow mdframed navigator needspace tablefootnote tocloft xpatch multirow bbding mathastext isomath relsize")
-  system("tlmgr update --all")
-  system("sudo apt-get update")
-  system("sudo apt-get install texlive-bibtex-extra")
+  if (!file.exists("travis_trattex_built") ||
+      as.double(difftime(Sys.time(), file.mtime("travis_trattex_built"), units = "days")) > 30) {
+    system("source ./travis/texlive.sh")
+    system("tlmgr install acronym bigfoot blindtext chngcntr cmap nowidow mdframed navigator needspace tablefootnote tocloft xpatch multirow bbding mathastext isomath relsize")
+    system("tlmgr update --all")
+    system("sudo apt-get update")
+    system("sudo apt-get install texlive-bibtex-extra")
+    system('wget "https://sourceforge.net/projects/biblatex-biber/files/biblatex-biber/current/binaries/Linux/biber-linux_x86_64.tar.gz"')
+    system('tar xzf biber-linux_x86_64.tar.gz')
+    system('export PATH=$PATH:$PWD')
+    system('tlmgr update biber')
+    file.create("travis_grattex_built")
+  }
 }
